http types file because http.h is a behemoth readability?
